fix: fix engine not starting if move played too quickly on page load
also cached the engine lines of the starting position so that moves can be added at the beginning and be classified immediately without having to redundantly evaluate the starting position. this does obviously NOT work with different starting positions, but if you are able to put a pgn in, the engine will have loaded to evaluate the initial position.