feat(fe): reclassify moves if engine depth setting increased
if there is already a classification on a move and the previous one does not have sufficient lines for a classification, the insufficient lines error message is silent. it is only if the current move does not have a classification that this error needs to be shown - otherwise, the program can silently remain on the current classification.