fix(sh): consider trapped piece and normal danger levels with different equality
to check if a piece is trapped by danger levels, it needs to be verified that the trapped piece moving is the thing that causes a more pressing danger to exist. in normal danger levels, like a counterattacking situation, you don't need this check. distinction had to be made to fix issues with very particular Chess positions where pieces are incorrectly considered trapped because moving them leaves an unrelated piece hanging.